Factor out CheatNavController from WaypointRobotController, add a rate limit.
Consider making separate nav, guidance, and control classes, so that
they can be arbitrarily combined...
